problem ze srednia klasy - propozycja wstepna rozwiazania - petla

+++ Dziennik TDD/Dziennik.hpp
@@ -8,7 +8,7 @@ public:
 	bool isEmpty() const;
 	void addStudent(const Student& s);
 	void deleteStudent(const Student& s);
-	void addGrade(Student& s, int grade);
+	void addGrade(Student& s, double grade);
 	double returnAverage(const Student& s);
 	double returnClassAverage();
 	std::vector<double> returnGrades(const Student& s);
